requirements.txt
gradio
pandas
python-dotenv
langchain-community
langchain-core
langchain-google-genai
pydantic
pypdf
docx2txt
openpyxl
unstructured
markdown
